Trying to fix some stuff in the mission cleanup
If the hidden cave segments are never revealed, then some tokens are stuck inside the explore bag. Trying to get them out.

Current issue is that by taking them out of the bag and placing them in the cleanup zone, the travel time of these objects is too slow - the rest of the algorithm runs before those objects reach the table